requirements.txt
Flask==2.0.3
python-dotenv==0.19.2
pycryptodome==3.10.1